Remove versioneer
Versioneer needed some additional files like _version.py versioneer.py.
This additional files in a unpatched version broke the linting with
flake8 and the type checking.

We set the __version__ variable to 0.0.0 to be ready for the
switch to the new tool Dynamic versioning plugin for Poetry
(https://github.com/mtkennerly/poetry-dynamic-versioning).